PCG.Command.Build

Description

Provides the types for the Build command along with a semantic definition to be consumed by the stream parser.

Documentation

data BuildCommand

The BUILD command specifying how a component graph should be constructed. output should be directed.

Constructors

BuildCommand Int ConstructionType

data ConstructionType

Different possible types of component graph construction.

Constructors

WagnerTree

WheelerNetwork

WheelerForest

buildCommandSpecification :: CommandSpecification BuildCommand

Defines the semantics of interpreting a valid "Report" command from the PCG scripting language syntax.