PCG.Command.Read

Description

Provides the types fot the Read command allong with a semantic definition to be consumed by the stream parser.

Documentation

data CustomAlphabetOptions

Options for custom alphabets. Not sure how these will be evaluation.

Constructors

Init3D Bool

Level Int (Either CustomAlphabetStrategy Tiebreaker)

Ties Tiebreaker

data CustomAlphabetStrategy

Strategy for alphabet symbols.

Constructors

First

Last

AtRandom

data FileSpecification

The specification for a file to be read.

Constructors

UnspecifiedFile (NonEmpty FilePath)

AminoAcidFile (NonEmpty FilePath)

NucleotideFile (NonEmpty FilePath)

AnnotatedFile (NonEmpty FilePath)

ChromosomeFile (NonEmpty FilePath)

GenomeFile (NonEmpty FilePath)

CustomAlphabetFile (NonEmpty FilePath) TcmReference [CustomAlphabetOptions]

PrealignedFile FileSpecification TcmReference

data FileSpecificationContent

The content of a file along with a possibly associated TCM file content.

Constructors

SpecContent

Fields

dataFiles :: [FileResult]

tcmFile :: Maybe FileResult

newtype ReadCommand

The Read command containing the files paths to be read.

Constructors

ReadCommand (NonEmpty FileSpecification)

type FileContent = Text

The content of a file.

type FileResult = (FilePath, FileContent)

The context of reading a file along with the path the content originated from.

type TcmReference = Maybe FilePath

An optional reference to a TCM file.

newtype Tiebreaker

Describes how ties are to be broken. In what context ties are occuring, I'm not sure.

Constructors

Tiebreaker CustomAlphabetStrategy

readCommandSpecification :: CommandSpecification ReadCommand

Defines the semantics of interpreting a valid "Read" command from the PCG scripting language syntax.