PCG.Command.Report

Description

Provides the types fot the Report command allong with a semantic definition to be consumed by the stream parser.

Documentation

data ReportCommand

The REPORT command specifying what information should be output and where the output should be directed.

Constructors

ReportCommand OutputFormat OutputTarget

data OutputFormat

How the output should be formatted

Constructors

CrossReferences [FilePath]

Data

DotFile

DynamicTable

ImpliedAlignmentCharacters

LikelihoodModel

Metadata

SearchStats

SequenceStats

Terminals

TreeCosts

TreeStats

XML

data OutputTarget

Where the output stream should be directed.

Constructors

OutputToStdout

OutputToFile FilePath FileWriteMethod

data FileWriteMethod

Defines the writing mode for how files should be written to disk.

Constructors

Append

Overwrite

reportCommandSpecification :: CommandSpecification ReportCommand

Defines the semantics of interpreting a valid "Report" command from the PCG scripting language syntax.