PCG.Syntax

Description

Exports all the PCG syntax stuff.

I think that eventually the Combinators module should *not* be export from this module, just the stream parser and command types.

Documentation

data Command

All the commands of the PCG scripting language.

Constructors

BUILD BuildCommand

READ ReadCommand

REPORT ReportCommand

newtype Computation

A non-empty list of Commands to be sequentially evaluated.

Constructors

Computation (NonEmpty Command)

computationalStreamParser :: (FoldCase (Tokens s), MonadParsec e s m, Token s ~ Char) => m Computation

Parse a series of PCG commands.

commandStreamParser :: (FoldCase (Tokens s), MonadParsec e s m, Token s ~ Char) => m Command

Parse a single, well defined PCG command.

module PCG.Syntax.Combinators