PCG.Syntax.Combinators

Description

Provides the syntactic combinators for specifying commands conforming to the PCG scripting language syntax. Allows for specifying commands which will be interpreted by the PCG scripting language syntax in a a context-free manner. Being context-free allows for permutations of command component as long as no two command components are ambiguously specified. Such ambiguity may lead to unexpected results when using the provided parser runSyntax.

Commands must be specified using an Applicative style so that a Free Applicative can be used as the syntactic & semantic parser.

Documentation

data CommandSpecification z

Specifcation for a command in the PCG scripting language syntax.

data SyntacticArgument z

Component of a semantic command embedded in the PCG scripting language syntax.

newtype ArgumentIdentifier

A identifier to disambiguate values of the same type.

Constructors

ArgId String

Primative Free Applicaitve constructors

bool :: Ap SyntacticArgument Bool

Define a boolean value as part of a command specification.

int :: Ap SyntacticArgument Int

Define a integer value as part of a command specification.

real :: Ap SyntacticArgument Double

Define a real value as part of a command specification.

Things that look like an integer can be captured as a real value by this combinator. This can lead to ambiguity between int and real. To avoid ambiguity, use argId to require a disambiguating prefix on one or more of the command components.

text :: Ap SyntacticArgument String

Define a textual value as part of a command specification.

time :: Ap SyntacticArgument DiffTime

Define a temporal value in minutes as part of a command specification.

value :: String -> Ap SyntacticArgument ()

Define a unique literal value as part of a command specification.

Syntactic Free Applicative constructors

argList :: Ap SyntacticArgument a -> Ap SyntacticArgument a

A list of arguments as part of a command.

command :: String -> Ap SyntacticArgument a -> CommandSpecification a

Specifies a command in the PCG scripting language defined by the command name and the argument structure.

choiceFrom :: Foldable f => f (Ap SyntacticArgument a) -> Ap SyntacticArgument a

Matches exactly one of the provided arguments to be used in the command.

argId :: String -> Ap SyntacticArgument a -> Ap SyntacticArgument a

Require a prefix on an agrument value to disambiguate it from other argument values.

argIds :: Foldable f => f String -> Ap SyntacticArgument a -> Ap SyntacticArgument a

Require a prefix on an agrument value to disambiguate it from other argument values. Accepts multiple aliases for the prefix used to disambiuate the argument.

manyOf :: Ap SyntacticArgument z -> Ap SyntacticArgument [z]

Produce zero or more of the provided argument for the command.

someOf :: Ap SyntacticArgument z -> Ap SyntacticArgument (NonEmpty z)

Produce one or more of the provided argument for the command.

withDefault :: Ap SyntacticArgument a -> a -> Ap SyntacticArgument a

Provide a default value for the argument if it is missing from the user input.

MonadParsec based syntactic interpreter

parseCommand :: (FoldCase (Tokens s), MonadParsec e s m, Token s ~ Char) => CommandSpecification a -> m a

Parse a command specification.

runSyntax :: (FoldCase (Tokens s), MonadParsec e s m, Token s ~ Char) => Ap SyntacticArgument a -> m a

Create a MonadParsec parser matching the specified semantics of the command.

Whitespace definition of the syntax

whitespace :: forall e s m. (MonadParsec e s m, Token s ~ Char) => m ()

Defines the whitespace for the syntax.

Exported for reuse by other MonadParsec based Free Monad interpreters.